shiping feature by shiproket API

+++ database/migrations/2025_07_19_080704_add_campaign_id_to_coin_campaigns_table.php
@@ -4,29 +4,23 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class AddCampaignIdToCoinCampaignsTable extends Migration
+return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     *
-     * @return void
-     */
     public function up()
     {
-        Schema::table('coin_campaigns', function (Blueprint $table) {
-            $table->string('campaign_id')->unique()->after('id'); 
-        });
+        if (!Schema::hasColumn('coin_campaigns', 'campaign_id')) {
+            Schema::table('coin_campaigns', function (Blueprint $table) {
+                $table->string('campaign_id')->after('id');
+            });
+        }
     }
 
-    /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
     public function down()
     {
-        Schema::table('coin_campaigns', function (Blueprint $table) {
-             $table->dropColumn('campaign_id');
-        });
+        if (Schema::hasColumn('coin_campaigns', 'campaign_id')) {
+            Schema::table('coin_campaigns', function (Blueprint $table) {
+                $table->dropColumn('campaign_id');
+            });
+        }
     }
-}
+};
